server/requirements.txt
# FastAPI 核心
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据验证
pydantic>=2.0.0

# HTTP 客户端
httpx>=0.25.0

# OpenAI 兼容 API（用于调用 ModelScope Qwen3）
openai>=1.0.0

# 其他工具
python-dotenv>=1.0.0
